remote-desktop-src: Only record frames at specified framerate
Drop frames if the paint signal came earlier than what the framerate
expects. This has some side effects; for example if the refresh rate of
the monitor is slightly lower than 60, we might end up dropping too
many frames, and we might end up not drawing the most recent frame, but
will be fixed later.