MetaScreen: add meta_screen_get_pixel_size to get physical screen size
When using logical screen coordinates, meta_screen_get_size returns the
logical size of the screen, but in some cases we might need to know
the actual physical size in pixels. Thus adding this value to the screen.

In MetaMonitorManager we have screen values as pixel size an logical_screen
sizes as logical, in MetaMonitor it's better to keep rect and pixel_rect
not to change this everywhere.