clutter/actor: Always ceil resource scale
Testing shows that ceiling the resource scale and always scaling down
pretty much always looks a lot better. The reason for this is most
likely that the widget rendered is not perfectly pixel aligned
(position and size given the paint scale) meaning even if we load an
almost perfectly sized image / texture, we'll still scale it ever so
slightly, and doing that is much more visible than just preparing a
larger resource, scaling it down.